+++ src/pages/home/CategoryList.tsx
@@ -2,7 +2,7 @@ import React, { useEffect, useState } from 'react';
 import { useSelector } from 'react-redux';
 import { selectActiveIndex } from '@stores/slices/NavCatrgorySlice.ts';
 import { useQuery } from 'react-query';
-import { getCommunityData } from '@apis/axiosInstance/axiosInstance';
+import { getCommunityData } from '@apis/community/community.ts';
 import * as S from '@pages/home/styles/CategoryListStyle';
 import { Community, QueryData } from '@interfaces/Community.ts';
 import CategorySearch from '@pages/home/CategorySearch.tsx';
